requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
clarifai==10.10.1
clarifai-grpc==10.10.1
clarifai-protocol==0.0.10
click==8.1.7
contextlib2==21.6.0
fsspec==2024.6.1
googleapis-common-protos==1.66.0
grpcio==1.64.3
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
inquirerpy==0.3.4
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==1.26.4
packaging==24.2
pfzy==0.3.4
pillow==11.0.0
prompt_toolkit==3.0.48
protobuf==5.27.3
Pygments==2.18.0
pytesseract==0.3.13
python-rapidjson==1.20
python-telegram-bot==21.7
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
schema==0.7.5
sniffio==1.3.1
tabulate==0.9.0
tqdm==4.67.0
tritonclient==2.51.0
urllib3==2.2.3
wcwidth==0.2.13
